backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pydantic==2.9.2
crewai==0.141.0
crewai-tools==0.51.1
langchain==0.3.20
langchain-community==0.3.19
duckduckgo-search==7.5.2
databricks-sdk==0.46.0
python-dotenv==1.0.0